Move playback logic into a hook and allow to specify the end
