+++ packages/client/package.json
@@ -61,7 +61,7 @@
     "react-refresh": "^0.9.0",
     "strict-event-emitter-types": "^2.0.0",
     "typescript": "4.9.5",
-    "webpack-dev-server": "^4.7.3"
+    "webpack-dev-server": "^4.15.1"
   },
   "dependencies": {
     "@datadog/browser-rum": "^3.6.12",
